+++ lib/pleroma/web/mastodon_api/views/conversation_view.ex
@@ -34,7 +34,11 @@ defmodule Pleroma.Web.MastodonAPI.ConversationView do
       id: participation.id |> to_string(),
       accounts: render(AccountView, "index.json", users: users, as: :user),
       unread: !participation.read,
-      last_status: render(StatusView, "show.json", activity: activity, for: user)
+      last_status:
+        render(StatusView, "show.json",
+          activity: activity,
+          direct_conversation_id: participation.id
+        )
     }
   end
 end